Moved functions out of `universe`
