+++ web/app/creators/maths.py
@@ -1,0 +1,10 @@
+import numpy as np
+from numpy import random as r
+
+
+def rnd(n, s):
+    """
+    returns positive int, within normal distribution
+    n = mean, s = std
+    """
+    return int(abs(np.ceil(r.normal(n, s))))
