+++ web/app/creators/universe.py
@@ -1,41 +1,19 @@
-import pickle
 import numpy as np
 import yaml
 from numpy import random as r
 from datetime import datetime
-import os
-print(os.listdir())
 
-# Depending on where this is run, it could be back one dir.
-try:
-    syllables = pickle.load(open("../data/syllables.p", "rb"))
-    pdata = yaml.safe_load(open('web/app/creators/specs/planet.yaml'))["planet_types"]
-except FileNotFoundError:
-    syllables = pickle.load(open("data/syllables.p", "rb"))
-    pdata = yaml.safe_load(open('web/app/creators/specs/planet.yaml'))["planet_types"]
-    mdata = yaml.safe_load(open('web/app/creators/specs/moon.yaml'))["moon_types"]
+from . import maths
+from . import language
+
+
+pdata = yaml.safe_load(open('web/app/creators/specs/planet.yaml'))["planet_types"]
+mdata = yaml.safe_load(open('web/app/creators/specs/moon.yaml'))["moon_types"]
 
 # TODO Get some stats on star types
 # sdata = {"radius_mean": 109, "radius_std": 1, "class": "G"}
 sdata = {"radius": 106, "class": "G"}
 
-
-
-
-#%%
-def make_word(n, spaces=True):
-    # TODO: Spaces not implemented
-    syl = np.random.choice(syllables, n)
-    word = "".join(syl)
-    return word.capitalize()
-
-
-def rnd(n, s):
-    """
-    returns positive int, within normal distribution
-    n = mean, s = std
-    """
-    return int(abs(np.ceil(r.normal(n, s))))
 
 
 def uuid(n=8):
@@ -44,30 +22,30 @@ def uuid(n=8):
 
 def sort_planets(t):
     if t == "terrestrial":
-        return rnd(0.39, 1.52)
+        return maths.rnd(0.39, 1.52)
     if t == "gas":
-        return rnd(5.2, 10)
+        return maths.rnd(5.2, 10)
     if t == "ice":
-        return rnd(15, 29)
+        return maths.rnd(15, 29)
     if t == "dwarf":
-        return rnd(30, 50)
+        return maths.rnd(30, 50)
 
 
 def make_star():
     star = sdata
-    star["name"] = make_word(rnd(2, 1))
+    star["name"] = language.make_word(maths.rnd(1, 1))
     star["objid"] = uuid(n=13)
     star["label"] = "star"
     return star
 
 
 def make_planet(t, orbiting):
-    planet = {"class": t, "name": make_word(rnd(2, 1))}
+    planet = {"class": t, "name": language.make_word(maths.rnd(2, 1))}
     planet["label"] = "planet"
     planet["objid"] = uuid(n=13)
     planet["mass"] = abs(r.normal(pdata[t]["mass_mean"], pdata[t]["mass_std"]))
     planet["radius"] = abs(r.normal(pdata[t]["radius_mean"], pdata[t]["radius_std"]))
-    planet["orbitsDistance"] = rnd(pdata[t]["distance_min"], pdata[t]["distance_max"])
+    planet["orbitsDistance"] = maths.rnd(pdata[t]["distance_min"], pdata[t]["distance_max"])
     planet["orbitsId"] = orbiting["objid"]
     planet["orbitsName"] = orbiting["name"]
     planet["isSupportsLife"] = False
@@ -82,7 +60,7 @@ def make_homeworld(orbiting, data):
     return planet
 
 def make_moon(t, planets):
-    moon = {"class": t, "name": make_word(rnd(2, 1))}
+    moon = {"class": t, "name": language.make_word(maths.rnd(2, 1))}
     moon["label"] = "moon"
     moon["objid"] = uuid(n=13)
     moon["mass"] = abs(r.normal(mdata[t]["mass_mean"], mdata[t]["mass_std"]))
@@ -104,7 +82,7 @@ def build_homeSystem(data, username):
     }
     systemid = uuid(n=13)
     system = {
-        "name": make_word(rnd(2, 1)),
+        "name": language.make_word(maths.rnd(2, 1)),
         "label": "system",
         "objid": systemid,
     }
