+++ web/app/views.py
@@ -5,6 +5,8 @@ from django.shortcuts import redirect, render
 from app.models import *
 
 from .creators import universe
+
+
 from .forms import HomeSystemForm, SignUpForm, QueryForm
